pyproject.toml
[project]
name = "pythonProject3"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name", email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]
packages = [
    { include = "pythonproject3", from = "src" }
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.3.0"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.17.1"
pytest = "^8.4.1"
pytest-cov = "^7.0.0"

[tool.black]
line-length = 119
target-version = ['py38']
skip-string-normalization = true

[tool.isort]
 profile = "black"
 line_length = 119
 known_third_party = []
 skip = ["__pycache__", ".egg-info", ".git", "venv"]
 force_single_line = true
 multi_line_output = 1

[tool.mypy]
files = "src"
ignore_missing_imports = true
warn_return_any = true
disallow_untyped_defs = true
exclude = "venv"